// generated/src/app/api-models/cap-rate-header.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapRateHeader {
  seqCapRateHdr: number;
  capModelId: string;
  capRateId: string;
  detValue1: string;
  detValue2: string;
  detValue3: string;
  effectiveDate: Date;
  endDate: Date;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}